Moved check into Authorizer code instead
(cherry picked from commit 4963383f2f54bf1e9d3ac489b2141f9f590e2a59)
